Add Prng, for consistent sampling. Use positive Long for cvr.sampleNum. ComparisonSampler needs contest to properly filter.

+++ core/src/main/kotlin/org/cryptobiotic/rlauxe/core/Cvr.kt
@@ -37,8 +37,8 @@ open class Cvr(
     }
 }
 
-// Mutable form of Cvr.
-class CvrUnderAudit(val cvr: Cvr, override val phantom: Boolean, var sampleNum: Int = 0): CvrIF {
+/** Mutable version of Cvr. sampleNum >= 0  */
+class CvrUnderAudit(val cvr: Cvr, override val phantom: Boolean, var sampleNum: Long = 0L): CvrIF {
     override val id = cvr.id
     override val votes = cvr.votes
